Pairwise: add debug printout and fix get_num_candidates signature.
get_num_candidates would return a double which is... not the best.
Fixed to return size_t instead, which led to a bunch of follow-on
effects and lots of methods having to be changed to use size_t
instead of ints.

+++ src/pairwise/beatpath.h
@@ -40,8 +40,8 @@ class beatpath : public abstract_condmat {
 		beatpath(const election_t & scores, size_t num_candidates,
 			pairwise_type kind);
 
-		double get_num_candidates() const {
-			return (contents.size());
+		size_t get_num_candidates() const {
+			return contents.size();
 		}
 };
